requirements.txt
pandas>=2.0
pyyaml>=6.0
openpyxl>=3.1
streamlit>=1.37
numpy>=1.25
altair>=5.0
